move spawning player to a test mod
