+++ game/server_main.lua
@@ -8,18 +8,6 @@ function game.onPlayerJoin(callback)
     table.insert(game.playerJoinCallbacks, callback)
 end
 
-local function spawnPlayer(player)
-    local cs = game.world.CHUNK_SIZE
-    local ws = game.world.WORLD_SIZE
-    local x = (ws * cs) / 2
-    local z = x
-    local y = 0
-    while game.world.getBlock(x, y, z) ~= 0 and  game.world.getBlock(x, y + 1, z) ~= 0 do
-        y = y + 1
-    end
-    player:setPosition(x, y + 3, z);
-end
-
 --Run callback functions
 function game.runPlayerJoinCallback(player)
     for _, callback in ipairs(game.playerJoinCallbacks) 
@@ -27,7 +15,5 @@ function game.runPlayerJoinCallback(player)
         callback(player)
     end
 end
-
-game.onPlayerJoin(spawnPlayer)
 
 dofile("game/voxel_types.lua")
